+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.0] - 2024-XX-XX
+## [0.4.0] - 2025-10-31
 ### BREAKING CHANGES: Unified Motion Store API
 
 This release unifies the motion store API for better ergonomics and performance.